legacy: Accept pci id generalizations
While diff is usually sufficient, there are perfectly valid
changes such as:

from cpu:type:x86,ven0000fam0006mod0066:feature:*0083*
to   cpu:type:x86,ven0000fam0006mod0066:feature:*

As these are generalizations, we do not wish to yield a failure.
To remedy that, we check whether or not the old one, as a string,
is satisfied by the new one, as a regular expression.

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>