serge-testsuite: replace test script with a wrapper (#446)
We recently moved the beakerlib wrapper for the upstream
selinux-testuite to a public repo, which should serve as a common source
for all users of this test (currently RH internal repo and CKI).

This commit replaces the duplicated scripts with a thin wrapper that
simply downloads the current version of the beakerlib script from the
public repo and runs it.

This will make it easier for the SELinux team to maintain this test as
it will now be in a single place and CKI will benefit from all
fixes/improvements without the need to send pull requests against the
tests-beaker repo.

Note that the requirements are now removed from the Makefile since the
test downloads them on its own now (to avoid the need to update each
wrapper even when a new requirement is added). The test already needed a
universal package install function to install the right version of
certain kernel subpackages so this doesn't really add much complexity to
it.

The test also no longer depends on the selinux-policy/common library (it
only used a tiny subset of it, which was easy to open-code), so its copy
is also removed.

One caveat is that the test will no longer work on RHEL-5, since git
clone on longer works there with src.fedoraproject.org (SSL/TLS
compatibility issues). However, considering the current support level of
RHEL-5, this is not a big loss.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>